requirements.txt
fastapi
tarantool
python-dotenv
pytest
gunicorn
uvicorn